My service offering catalogue for cloud and disruptive technologies
This is an initial series of documents that I have drafted as part of my
offerings in areas of AI/ML, Cloud, data center, automation, digital
platform, R&D services, engineering services, with a focus for MS Azure
cloud platform.  These documents are service catalogue that can be used
for engagements and proper offerings to clients. This demonstrated that
we have matured offerings for the cloud and can offer a decent service
for our clients who seek to migrate to cloud platform